# docs/contracts/public-rest.openapi.yaml
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/auth/sessions/refresh": {
      "post": {
        "operationId": "Auth.refresh",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": false,
                "required": [
                  "refreshToken"
                ],
                "properties": {
                  "refreshToken": {
                    "type": "string",
                    "writeOnly": true,
                    "maxLength": 512
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthSession"
                }
              }
            }
          },
          "default": {
            "description": "Stable non-sensitive error envelope.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        },
        "summary": "Rotate a refresh session and issue fresh tokens",
        "tags": [
          "authentication"
        ],
        "x-swar-auth-kind": "PUBLIC",
        "x-swar-required-permissions": [],
        "x-swar-rate-limit-category": "SENSITIVE",
        "x-swar-idempotency": "NOT_APPLICABLE"
      }
    },
...
    "/api/v1/risk-policies/active": {
      "get": {
        "operationId": "RiskPolicy.active",
        "parameters": [
          {
            "name": "policyKey",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RiskPolicy"
                }
              }
            }
          },
          "default": {
            "description": "Stable non-sensitive error envelope.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Retrieve the active immutable policy version",
        "tags": [
          "risk-policy"
        ],
        "x-swar-auth-kind": "USER_ACCESS_JWT",
        "x-swar-required-permissions": [
          "risk-policy.read"
        ],
        "x-swar-rate-limit-category": "QUERY",
        "x-swar-idempotency": "NOT_APPLICABLE"
      }
    },
    "/api/v1/risk-policies/{policyKey}": {
      "put": {
        "operationId": "RiskPolicy.put",
        "parameters": [
          {
            "name": "policyKey",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PutRiskPolicyDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RiskPolicy"
                }
              }
            }
          },
          "default": {
            "description": "Stable non-sensitive error envelope.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create and optionally activate an immutable risk-policy version",
        "tags": [
          "risk-policy"
        ],
        "x-swar-auth-kind": "USER_ACCESS_JWT",
        "x-swar-required-permissions": [
          "risk-policy.manage"
        ],
        "x-swar-rate-limit-category": "SENSITIVE",
        "x-swar-idempotency": "IDEMPOTENCY_KEY_REQUIRED"
      }
    },
...
    "/api/v1/interventions/{interventionId}/verification-challenges": {
      "post": {
        "operationId": "Interventions.request",
        "parameters": [
          {
            "name": "interventionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateVerificationChallengeDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VerificationChallenge"
                }
              }
            }
          },
          "default": {
            "description": "Stable non-sensitive error envelope.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Request independent step-up or callback verification",
        "tags": [
          "interventions"
        ],
        "x-swar-auth-kind": "USER_ACCESS_JWT",
        "x-swar-required-permissions": [
          "intervention.resolve"
        ],
        "x-swar-rate-limit-category": "SENSITIVE",
        "x-swar-idempotency": "IDEMPOTENCY_KEY_REQUIRED"
      }
    },
    "/api/v1/internal/verifications/{challengeId}/result": {
      "post": {
        "operationId": "Interventions.complete",
        "parameters": [
          {
            "name": "challengeId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-SWAR-Service",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "swar-verifier"
              ]
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompleteVerificationChallengeDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VerificationResult"
                }
              }
            }
          },
          "default": {
            "description": "Stable non-sensitive error envelope.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "verifierService": []
          }
        ],
        "summary": "Record an authenticated independent-verifier adapter result",
        "tags": [
          "interventions"
        ],
        "x-swar-auth-kind": "VERIFIER_SERVICE",
        "x-swar-required-permissions": [],
        "x-swar-rate-limit-category": "SENSITIVE",
        "x-swar-idempotency": "NOT_APPLICABLE"
      }
    },
    "/api/v1/interventions/{interventionId}/release": {
      "post": {
        "operationId": "Interventions.release",
        "parameters": [
          {
            "name": "interventionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReleaseInterventionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatusResult"
                }
              }
            }
          },
          "default": {
            "description": "Stable non-sensitive error envelope.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Release a hold only after a persisted PASSED independent challenge",
        "tags": [
          "interventions"
        ],
        "x-swar-auth-kind": "USER_ACCESS_JWT",
        "x-swar-required-permissions": [
          "intervention.resolve"
        ],
        "x-swar-rate-limit-category": "SENSITIVE",
        "x-swar-idempotency": "IDEMPOTENCY_KEY_REQUIRED"
      }
    }
  },
  "info": {
    "title": "SWAR REST and callback API",
    "description": "Versioned contracts for authenticated users, internal services, and signed LiveKit webhooks. Examples are fictional.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "/api/v1",
      "description": "Versioned SWAR API base path"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "mlService": {
        "type": "http",
        "scheme": "bearer"
      },
      "verifierService": {
        "type": "http",
        "scheme": "bearer"
      },
      "liveKitWebhook": {
        "type": "apiKey",
        "in": "header",
        "name": "Authorization"
      }
    },
    "schemas": {
...
      "PutRiskPolicyDto": {
        "type": "object",
        "properties": {
          "version": {
            "type": "string",
            "maxLength": 40
          },
          "schemaVersion": {
            "type": "string",
            "maxLength": 40,
            "example": "1.0.0"
          },
          "policyDocument": {
            "type": "object",
            "additionalProperties": true
          },
          "activate": {
            "type": "boolean",
            "description": "Activate this immutable version after creation."
          }
        },
        "required": [
          "version",
          "schemaVersion",
          "policyDocument",
          "activate"
        ]
      },
      "CreateVerificationChallengeDto": {
        "type": "object",
        "properties": {
          "method": {
            "type": "string",
            "enum": [
              "CALLBACK",
              "OUT_OF_BAND"
            ]
          }
        },
        "required": [
          "method"
        ]
      },
      "CompleteVerificationChallengeDto": {
        "type": "object",
        "properties": {
          "organizationId": {
            "type": "string",
            "format": "uuid"
          },
          "result": {
            "type": "string",
            "enum": [
              "PASSED",
              "FAILED"
            ]
          },
          "resultCode": {
            "type": "string",
            "maxLength": 80
          }
        },
        "required": [
          "organizationId",
          "result",
          "resultCode"
        ]
      },
      "ReleaseInterventionDto": {
        "type": "object",
        "properties": {
          "verificationChallengeId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "required": [
          "verificationChallengeId"
        ]
      },
      "ErrorEnvelope": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "code",
          "message",
          "requestId"
        ],
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "requestId": {
            "type": "string"
          },
          "details": {
            "type": "object",
            "additionalProperties": true
          }
        }
      },
      "AuthSession": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "accessToken",
          "refreshToken",
          "tokenType",
          "expiresIn",
          "principal"
        ],
        "properties": {
          "accessToken": {
            "type": "string",
            "readOnly": true,
            "x-swar-sensitive": true
          },
          "refreshToken": {
            "type": "string",
            "readOnly": true,
            "x-swar-sensitive": true
          },
          "tokenType": {
            "type": "string",
            "enum": [
              "Bearer"
            ]
          },
          "expiresIn": {
            "type": "integer",
            "minimum": 1
          },
          "principal": {
            "type": "object",
            "additionalProperties": false,
            "required": [
              "userId",
              "membershipId",
              "organizationId",
              "deviceId",
              "roles"
            ],
            "properties": {
              "userId": {
                "type": "string",
                "format": "uuid"
              },
              "membershipId": {
                "type": "string",
                "format": "uuid"
              },
              "organizationId": {
                "type": "string",
                "format": "uuid"
              },
              "deviceId": {
                "type": "string",
                "format": "uuid"
              },
              "roles": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
...
      "StatusResult": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "consentId": {
            "type": "string",
            "format": "uuid"
          },
          "voiceprintId": {
            "type": "string",
            "format": "uuid"
          },
          "interventionId": {
            "type": "string",
            "format": "uuid"
          },
          "status": {
            "type": "string"
          }
        },
        "required": [
          "status"
        ]
      },
...
      "RiskPolicy": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "riskPolicyId",
          "policyKey",
          "version",
          "schemaVersion",
          "policyDocument",
          "status",
          "effectiveAt"
        ],
        "properties": {
          "riskPolicyId": {
            "type": "string",
            "format": "uuid"
          },
          "policyKey": {
            "type": "string"
          },
          "version": {
            "type": "string"
          },
          "schemaVersion": {
            "type": "string"
          },
          "policyDocument": {
            "type": "object",
            "additionalProperties": true
          },
          "status": {
            "enum": [
              "DRAFT",
              "ACTIVE",
              "RETIRED"
            ]
          },
          "effectiveAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        }
      },
      "VerificationChallenge": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "challengeId",
          "interventionId",
          "status",
          "method",
          "expiresAt"
        ],
        "properties": {
          "challengeId": {
            "type": "string",
            "format": "uuid"
          },
          "interventionId": {
            "type": "string",
            "format": "uuid"
          },
          "status": {
            "enum": [
              "PENDING",
              "PASSED",
              "FAILED",
              "EXPIRED",
              "CANCELLED"
            ]
          },
          "method": {
            "type": "string"
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time"
          },
          "resultCode": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "VerificationResult": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "challengeId",
          "status",
          "resultCode"
        ],
        "properties": {
          "challengeId": {
            "type": "string",
            "format": "uuid"
          },
          "status": {
            "enum": [
              "PASSED",
              "FAILED"
            ]
          },
          "resultCode": {
            "type": "string"
          }
        }
      }
    }
  }
}